requirements.txt
requests>=2.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pytest>=8.0.0
beautifulsoup4>=4.12.0
numpy>=1.24.0
faiss-cpu>=1.7.4